Release Travel Anywhere v2.0
New :
+ Bug fix data not sync with cloud
+ Bug fix crash when first time buy ticket
- Animation in PesanActivity
- More Secure Admin Manager
- New Dialog Permission for notification and Alarm permits
- Allow access without Alarm Granted

Feature:
- Authentication
- Semi Online Database
- Pesan Ticket Integrated with Database
- Full History
- Profile
- Admin Management

+++ app/src/main/java/com/ppb/travelanywhere/adapter/UsersManagerAdapter.kt
@@ -1,7 +1,5 @@
 package com.ppb.travelanywhere.adapter
 
-import android.content.Context
-import android.preference.Preference
 import android.view.LayoutInflater
 import android.view.ViewGroup
 import android.widget.Toast
@@ -32,6 +30,10 @@ class UsersManagerAdapter(
                             Toast.makeText(binding.root.context, "Tidak bisa mengubah role sendiri", Toast.LENGTH_SHORT).show()
                             return@launch
                         }
+                        if (user.role == "Role_Owner") {
+                            Toast.makeText(binding.root.context, "Tidak memiliki hak mengubah role owner", Toast.LENGTH_SHORT).show()
+                            return@launch
+                        }
                         val newRole = FireAuth().updateUserRole(user.id, user.role)
                         user.role = newRole
                         buttonRole.text = newRole.removePrefix("Role_").lowercase()
@@ -39,6 +41,14 @@ class UsersManagerAdapter(
                 }
                 buttonDeleteUser.setOnClickListener {
                     scope.launch {
+                        if (user.id == ApplicationPreferencesManager(binding.root.context).usernameId) {
+                            Toast.makeText(binding.root.context, "Mohon menghapus diri Anda melalui profil untuk memastikan", Toast.LENGTH_LONG).show()
+                            return@launch
+                        }
+                        if (user.role == "Role_Owner") {
+                            Toast.makeText(binding.root.context, "Tidak memiliki hak mengubah role owner", Toast.LENGTH_SHORT).show()
+                            return@launch
+                        }
                         val result = FireAuth().deleteUser(user.id)
                         if (result) {
                             val position = currentList.indexOf(user)
